add-reviews: fix: теперь оценки ревью хранятся в одной таблице
